requirements.txt
# python 3

gitpython
argparse
python-rpm-spec
